Cast PDO last insert ID result to int

+++ src/Pdo/PdoDatabase.php
@@ -65,6 +65,6 @@ class PdoDatabase extends PDO implements PdoDatabaseInterface
      */
     public function lastId(): int
     {
-        return parent::lastInsertId();
+        return (int) parent::lastInsertId();
     }
 }
